rag_chatbot/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.48.1
requests==2.32.5
openai==1.99.9
pypdf==5.9.0
faiss-cpu==1.13.2
numpy==2.3.2
tenacity==9.1.2
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
